Added folders to group Recipe, Pantry, and Shopping List files together
