+++ CIS560-RecipeManager/CIS560-RecipeManager/CIS560-RecipeManager.csproj
@@ -62,39 +62,39 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Ingredient.cs" />
-    <Compile Include="IQuery.cs" />
-    <Compile Include="RecipeCategory.cs" />
+    <Compile Include="PantryManager\Ingredient.cs" />
+    <Compile Include="Repository\IQuery.cs" />
+    <Compile Include="RecipeManager\RecipeCategory.cs" />
     <Compile Include="RecipeDatabaseDataSet.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
       <DependentUpon>RecipeDatabaseDataSet.xsd</DependentUpon>
     </Compile>
-    <Compile Include="ShoppingList.cs" />
-    <Compile Include="uiAddNewRecipe_BASIC.cs">
+    <Compile Include="ShoppingListManager\ShoppingList.cs" />
+    <Compile Include="RecipeManager\uiAddNewRecipe_BASIC.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="uiAddNewRecipe_BASIC.Designer.cs">
+    <Compile Include="RecipeManager\uiAddNewRecipe_BASIC.Designer.cs">
       <DependentUpon>uiAddNewRecipe_BASIC.cs</DependentUpon>
     </Compile>
-    <Compile Include="IRecipeRepository.cs" />
-    <Compile Include="Recipe.cs" />
+    <Compile Include="Repository\IRecipeRepository.cs" />
+    <Compile Include="RecipeManager\Recipe.cs" />
     <Compile Include="RecipeDatabaseDataSet_DEMO.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
       <DependentUpon>RecipeDatabaseDataSet_DEMO.xsd</DependentUpon>
     </Compile>
-    <Compile Include="RecipeRepository.cs" />
-    <Compile Include="uiRecipeManager.cs">
+    <Compile Include="Repository\RecipeRepository.cs" />
+    <Compile Include="RecipeManager\uiRecipeManager.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="uiHome.Designer.cs">
+    <Compile Include="RecipeManager\uiHome.Designer.cs">
       <DependentUpon>uiRecipeManager.cs</DependentUpon>
     </Compile>
-    <Compile Include="uiPantryManager.cs">
+    <Compile Include="PantryManager\uiPantryManager.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="uiPantryManager.Designer.cs">
+    <Compile Include="PantryManager\uiPantryManager.Designer.cs">
       <DependentUpon>uiPantryManager.cs</DependentUpon>
     </Compile>
     <Compile Include="uiSplashWelcome.cs">
@@ -105,7 +105,7 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="uiAddNewRecipe_BASIC.resx">
+    <EmbeddedResource Include="RecipeManager\uiAddNewRecipe_BASIC.resx">
       <DependentUpon>uiAddNewRecipe_BASIC.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
@@ -117,10 +117,10 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="uiHome.resx">
+    <EmbeddedResource Include="RecipeManager\uiHome.resx">
       <DependentUpon>uiRecipeManager.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="uiPantryManager.resx">
+    <EmbeddedResource Include="PantryManager\uiPantryManager.resx">
       <DependentUpon>uiPantryManager.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="uiSplashWelcome.resx">
