--- CIS560-RecipeManager/CIS560-RecipeManager/Ingredient.cs
@@ -1,20 +1,0 @@
-﻿namespace CIS560_RecipeManager
-{
-    public class Ingredient
-    {
-        public int Id { get; }
-
-        public string Name { get; }
-
-        public int PantryQuantity { get; set; }
-
-        public string Unit { get; }
-
-        public Ingredient(int id, string name, string unit)
-        {
-            Id = id;
-            Name = name;
-            Unit = unit;
-        }
-    }
-}
