+++ CIS560-RecipeManager/CIS560-RecipeManager/RecipeManager/RecipeCategory.cs
@@ -1,0 +1,14 @@
+﻿namespace CIS560_RecipeManager
+{
+    public class RecipeCategory
+    {
+        public int Id { get; }
+        public string Name { get; }
+
+        public RecipeCategory(int id, string name)
+        {
+            Id = id;
+            Name = name;
+        }
+    }
+}
