--- CIS560-RecipeManager/CIS560-RecipeManager/ShoppingList.cs
@@ -1,15 +1,0 @@
-﻿namespace CIS560_RecipeManager
-{
-    public class ShoppingList
-    {
-        public int Id { get; }
-
-        public string Name { get; }
-
-        public ShoppingList(int id, string name)
-        {
-            Id = id;
-            Name = name;
-        }
-    }
-}
